+++ ontimize/components/table/extensions/header/table-button/o-table-button.component.ts
@@ -1,8 +1,11 @@
-import { Component, OnInit, Inject, forwardRef, EventEmitter, Injector, ViewEncapsulation } from '@angular/core';
+import { Component, OnInit, Inject, forwardRef, EventEmitter, Injector, ViewEncapsulation, ElementRef } from '@angular/core';
 import { Util, Codes } from '../../../../../utils';
+import { InputConverter } from '../../../../../decorators';
 import { OTableComponent } from '../../../o-table.component';
 
 export const DEFAULT_INPUTS_O_TABLE_BUTTON = [
+  'oattr: attr',
+  'enabled',
   'icon',
   'svgIcon: svg-icon',
   'iconPosition: icon-position',
@@ -31,15 +34,19 @@ export class OTableButtonComponent implements OnInit {
   public static DEFAULT_INPUTS_O_TABLE_BUTTON = DEFAULT_INPUTS_O_TABLE_BUTTON;
   public static DEFAULT_OUTPUTS_O_TABLE_BUTTON = DEFAULT_OUTPUTS_O_TABLE_BUTTON;
 
-  public onClick: EventEmitter<Object> = new EventEmitter<Object>();
+  onClick: EventEmitter<Object> = new EventEmitter<Object>();
 
-  public icon: string;
-  public svgIcon: string;
-  public olabel: string;
-  public iconPosition: string;
+  oattr: string;
+  @InputConverter()
+  enabled: boolean = true;
+  icon: string;
+  svgIcon: string;
+  olabel: string;
+  iconPosition: string;
 
   constructor(
     protected injector: Injector,
+    public elRef: ElementRef,
     @Inject(forwardRef(() => OTableComponent)) protected _table: OTableComponent
   ) {