+++ ontimize/components/table/o-table.component.ts
@@ -13,8 +13,9 @@ import { InputConverter } from '../../decorators';
 import { OTableDataSource } from './o-table.datasource';
 import { OFormComponent } from '../form/o-form.component';
 import { Codes, ObservableWrapper, Util } from '../../utils';
+import { PermissionsUtils } from '../../util/permissions';
 import { OServiceComponent } from '../o-service-component.class';
-import { OntimizeService, SnackBarService, OPermissions, OTablePermissions, PermissionsService } from '../../services';
+import { OntimizeService, SnackBarService, OPermissions, OTablePermissions, OTableMenuPermissions } from '../../services';
 import { OTableRowDirective } from './extensions/row/o-table-row.directive';
 import { OColumnTooltip } from './column/o-table-column.component';
 import { OTableStorage } from './extensions/o-table-storage.class';
@@ -36,7 +37,10 @@ import {
   OTableQuickfilterComponent,
   OTableEditableRowComponent,
   IColumnValueFilter,
-  ColumnValueFilterOperator
+  ColumnValueFilterOperator,
+  OTableMenuComponent,
+  OTableButtonsComponent,
+  OTableButtonComponent
 } from './extensions/header/o-table-header-components';
 
 import {
@@ -49,20 +53,12 @@ import {
 
 import {
   O_TABLE_DIALOGS,
-  OTableExportConfiguration,
-  OTableExportDialogComponent,
-  OTableLoadFilterDialogComponent,
-  OTableStoreFilterDialogComponent,
-  OTableVisibleColumnsDialogComponent,
-  OTableFilterByColumnDataDialogComponent,
-  OTableStoreConfigurationDialogComponent,
-  OTableApplyConfigurationDialogComponent
+  OTableFilterByColumnDataDialogComponent
 } from './extensions/dialog/o-table-dialog-components';
 
 import {
   O_TABLE_CELL_RENDERERS,
-  OBaseTableCellRenderer,
-  OTableCellRendererImageComponent
+  OBaseTableCellRenderer
 } from './column/cell-renderer/cell-renderer';
 
 import { O_TABLE_CELL_EDITORS } from './column/cell-editor/cell-editor';
@@ -301,9 +297,6 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   public paginator: OTablePaginatorComponent;
   @ViewChild(MatPaginator) matpaginator: MatPaginator;
   @ViewChild(OMatSort) sort: OMatSort;
-  @ViewChild('columnFilterOption') columnFilterOption: OTableOptionComponent;
-  @ContentChildren(OTableOptionComponent) tableOptions: QueryList<OTableOptionComponent>;
-  @ViewChild('menu') matMenu: MatMenu;
   @ViewChild(OTableEditableRowComponent) oTableEditableRow: OTableEditableRowComponent;
 
   // only for insideTabBugWorkaround
@@ -382,6 +375,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
 
   /*parsed inputs variables */
   protected _visibleColArray: Array<string> = [];
+
+  get originalVisibleColumns(): string {
+    return this.visibleColumns;
+  }
 
   get visibleColArray(): Array<any> {
     return this._visibleColArray;
@@ -477,7 +474,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   public onUpdateScrolledState: EventEmitter<any> = new EventEmitter();
   public rowWidth;
 
-  protected oTableStorage: OTableStorage;
+  oTableStorage: OTableStorage;
   storePaginationState: boolean = false;
 
   /* In the case the table havent paginationControl and pageable, the table has pagination virtual*/
@@ -501,6 +498,17 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   protected permissions: OTablePermissions;
+  matMenu: MatMenu;
+
+  @ViewChild('tableMenu')
+  oTableMenu: OTableMenuComponent;
+  @ContentChildren(OTableOptionComponent)
+  tableOptions: QueryList<OTableOptionComponent>;
+
+  @ViewChild('tableButtons')
+  oTableButtons: OTableButtonsComponent;
+  @ContentChildren(OTableButtonComponent)
+  tableButtons: QueryList<OTableButtonComponent>;
 
   constructor(
     injector: Injector,
@@ -527,31 +535,53 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   ngAfterViewInit() {
     this.afterViewInit();
     this.initTableAfterViewInit();
-    this.parsePermissions();
+    if (this.oTableMenu) {
+      this.matMenu = this.oTableMenu.matMenu;
+      this.oTableMenu.registerOptions(this.tableOptions.toArray());
+    }
+    if (this.oTableButtons) {
+      this.oTableButtons.registerButtons(this.tableButtons.toArray());
+    }
   }
 
   ngOnDestroy() {
     this.destroy();
   }
 
-  protected parsePermissions() {
-    // const self = this;
-    // this.permissions.actions.forEach((permission: OPermissions) => {
-
-    // });
+  getActionsPermissions(): OPermissions[] {
+    return this.permissions ? (this.permissions.actions || []) : [];
   }
 
-  protected getOColumnPermissions(attr: string): OPermissions {
-    const columns = this.permissions ? this.permissions.columns : [];
+  getMenuPermissions(): OTableMenuPermissions {
+    let result: OTableMenuPermissions = this.permissions ? this.permissions.menu : undefined;
+    return result ? result : {
+      visible: true,
+      enabled: true,
+      items: []
+    };
+  }
+
+  getOColumnPermissions(attr: string): OPermissions {
+    const columns = this.permissions ? (this.permissions.columns || []) : [];
     return columns.find(comp => comp.attr === attr) || { attr: attr, enabled: true, visible: true };
   }
 
-  protected checkEnabledActionPermission(attr) {
-    const actionsPerm = this.permissions.actions;
-    const permissions: OPermissions = (actionsPerm || []).find(p => p.attr === attr);
-    let enabledPermision = PermissionsService.checkEnabledPermission(permissions);
+  protected getActionPermissions(attr: string): OPermissions {
+    const actionsPerm = this.permissions ? (this.permissions.actions || []) : [];
+    const permissions: OPermissions = actionsPerm.find(p => p.attr === attr);
+    return permissions || {
+      attr: attr,
+      visible: true,
+      enabled: true
+    };
+  }
+
+  protected checkEnabledActionPermission(attr: string) {
+    const actionsPerm = this.permissions ? (this.permissions.actions || []) : [];
+    const permissions: OPermissions = actionsPerm.find(p => p.attr === attr);
+    let enabledPermision = PermissionsUtils.checkEnabledPermission(permissions);
     if (!enabledPermision) {
-      this.snackBarService.open(PermissionsService.MESSAGE_OPERATION_NOT_ALLOWED_PERMISSION);
+      this.snackBarService.open(PermissionsUtils.MESSAGE_OPERATION_NOT_ALLOWED_PERMISSION);
     }
     return enabledPermision;
   }
@@ -1148,47 +1178,15 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   onExportButtonClicked() {
-    let exportCnfg: OTableExportConfiguration = new OTableExportConfiguration();
-    // Table data
-    exportCnfg.data = this.getRenderedValue();
-    // get column's attr whose renderer is OTableCellRendererImageComponent
-    let colsNotIncluded: string[] = this._oTableOptions.columns.filter(c => void 0 !== c.renderer && c.renderer instanceof OTableCellRendererImageComponent).map(c => c.attr);
-    colsNotIncluded.forEach(attr => exportCnfg.data.forEach(row => delete row[attr]));
-    // Table columns
-    exportCnfg.columns = this._oTableOptions.visibleColumns.filter(c => colsNotIncluded.indexOf(c) === -1);
-    // Table column names
-    let tableColumnNames = {};
-    this._oTableOptions.visibleColumns.filter(c => colsNotIncluded.indexOf(c) === -1).map(c => tableColumnNames[c] = this.translateService.get(c));
-    exportCnfg.columnNames = tableColumnNames;
-    // Table column sqlTypes
-    exportCnfg.sqlTypes = this.getSqlTypes();
-    // Table service, needed for configuring ontimize export service with table service configuration
-    exportCnfg.service = this.service;
-
-    let dialogRef = this.dialog.open(OTableExportDialogComponent, {
-      data: exportCnfg,
-      disableClose: true
-    });
-
-    dialogRef.afterClosed().subscribe(result => result ? this.snackBarService.open('MESSAGES.SUCCESS_EXPORT_TABLE_DATA', { icon: 'check_circle' }) : null);
+    if (this.oTableMenu) {
+      this.oTableMenu.onExportButtonClicked();
+    }
   }
 
   onChangeColumnsVisibilityClicked() {
-    let dialogRef = this.dialog.open(OTableVisibleColumnsDialogComponent, {
-      data: {
-        originalVisibleColumns: Util.parseArray(this.visibleColumns, true),
-        columnsData: this._oTableOptions.columns
-      },
-      disableClose: true
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.visibleColArray = dialogRef.componentInstance.getVisibleColumns();
-        let columnsOrder = dialogRef.componentInstance.getColumnsOrder();
-        this._oTableOptions.columns.sort((a: OColumn, b: OColumn) => columnsOrder.indexOf(a.attr) - columnsOrder.indexOf(b.attr));
-      }
-    });
+    if (this.oTableMenu) {
+      this.oTableMenu.onChangeColumnsVisibilityClicked();
+    }
   }
 
   onMatTableContentChanged() {
@@ -1196,10 +1194,16 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   add() {
+    if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_INSERT)) {
+      return;
+    }
     super.insertDetail();
   }
 
   remove(clearSelectedItems: boolean = false) {
+    if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_DELETE)) {
+      return;
+    }
     let selectedItems = this.getSelectedItems();
     if (selectedItems.length > 0) {
       this.dialogService.confirm('CONFIRM', 'MESSAGES.CONFIRM_DELETE').then(res => {
@@ -1207,7 +1211,6 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
           if (this.dataService && (this.deleteMethod in this.dataService) && this.entity && (this.keysArray.length > 0)) {
             let filters = ServiceUtils.getArrayProperties(selectedItems, this.keysArray);
             this.daoTable.removeQuery(filters).subscribe(res => {
-              console.log('[OTable.remove]: response', res);
               ObservableWrapper.callEmit(this.onRowDeleted, selectedItems);
             }, error => {
               this.showDialogError(error, 'MESSAGES.ERROR_DELETE');
@@ -1249,6 +1252,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   reloadData() {
+    if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_REFRESH)) {
+      return;
+    }
+
     this.clearSelection();
     this.finishQuerySubscription = false;
     this.pendingQuery = true;
@@ -1385,6 +1392,12 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     if (event && column.editing && this.editingCell === event.currentTarget) {
       return;
     }
+    const columnPermissions: OPermissions = this.getOColumnPermissions(column.attr);
+    if (columnPermissions.enabled === false) {
+      console.warn(`${column.attr} edition not allowed due to permissions`);
+      return;
+    }
+
     this.clearSelectionAndEditing();
     this.selectedRow(row);
     this.editingCell = event.currentTarget;
@@ -1399,9 +1412,12 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   updateCellData(column: OColumn, data: any, saveChanges: boolean) {
-    // if (!this.checkEnabledActionPermission(PermissionsService.PERMISSIONS_ACTIONS_UPDATE_FORM)) {
-    //   return;
-    // }
+    if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_UPDATE)) {
+      const res = new Observable(innerObserver => {
+        innerObserver.error();
+      });
+      return res;
+    }
     column.editing = false;
     this.editingCell = undefined;
     if (saveChanges && this.editingRow !== undefined) {
@@ -1433,11 +1449,12 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   onShowsSelects(event?: any) {
-    this._oTableOptions.selectColumn.visible = !this._oTableOptions.selectColumn.visible;
-    this.updateSelectionColumnState();
+    if (this.oTableMenu) {
+      this.oTableMenu.onShowsSelects(event);
+    }
   }
 
-  protected updateSelectionColumnState() {
+  updateSelectionColumnState() {
     if (!this._oTableOptions.selectColumn.visible) {
       this.clearSelection();
     }
@@ -1566,69 +1583,34 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   onFilterByColumnClicked() {
-    if (this.showFilterByColumnIcon && this.dataSource.isColumnValueFilterActive()) {
-      const self = this;
-      this.dialogService.confirm('CONFIRM', 'MESSAGES.CONFIRM_DISCARD_FILTER_BY_COLUMN').then(res => {
-        if (res) {
-          self.dataSource.clearColumnFilters();
-        }
-        self.showFilterByColumnIcon = !res;
-      });
-    } else {
-      this.showFilterByColumnIcon = !this.showFilterByColumnIcon;
+    if (this.oTableMenu) {
+      this.oTableMenu.onFilterByColumnClicked();
     }
   }
 
-  onStoreFilterClicked(): void {
-    let dialogRef = this.dialog.open(OTableStoreFilterDialogComponent, {
-      data: this.oTableStorage.getStoredFilters().map(filter => filter.name),
-      width: '30vw',
-      disableClose: true
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.oTableStorage.storeFilter(dialogRef.componentInstance.getFilterAttributes());
-      }
-    });
+  onStoreFilterClicked() {
+    if (this.oTableMenu) {
+      this.oTableMenu.onStoreFilterClicked();
+    }
   }
 
-  onLoadFilterClicked(): void {
-    let dialogRef = this.dialog.open(OTableLoadFilterDialogComponent, {
-      data: this.oTableStorage.getStoredFilters(),
-      width: '30vw',
-      disableClose: true
-    });
-
-    dialogRef.componentInstance.onDelete.subscribe(filterName => this.oTableStorage.deleteStoredFilter(filterName));
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        let selectedFilterName: string = dialogRef.componentInstance.getSelectedFilterName();
-        if (selectedFilterName) {
-          let storedFilter = this.oTableStorage.getStoredFilterConf(selectedFilterName);
-          if (storedFilter) {
-            this.setFiltersConfiguration(storedFilter);
-            this.reloadPaginatedDataFromStart();
-          }
-        }
-      }
-    });
+  onLoadFilterClicked() {
+    if (this.oTableMenu) {
+      this.oTableMenu.onLoadFilterClicked();
+    }
   }
 
-  onClearFilterClicked(): void {
-    this.dialogService.confirm('CONFIRM', 'TABLE.DIALOG.CONFIRM_CLEAR_FILTER').then(result => {
-      if (result) {
-        this.clearFilters();
-        this.reloadPaginatedDataFromStart();
-      }
-    });
+  onClearFilterClicked() {
+    if (this.oTableMenu) {
+      this.oTableMenu.onClearFilterClicked();
+    }
   }
 
   clearFilters(triggerDatasourceUpdate: boolean = true): void {
     this.dataSource.clearColumnFilters(triggerDatasourceUpdate);
     this.showFilterByColumnIcon = false;
-    if (this.columnFilterOption) {
-      this.columnFilterOption.active = this.showFilterByColumnIcon;
+    if (this.oTableMenu && this.oTableMenu.columnFilterOption) {
+      this.oTableMenu.columnFilterOption.active = this.showFilterByColumnIcon;
     }
     if (this.oTableQuickFilterComponent) {
       this.oTableQuickFilterComponent.setValue(void 0);
@@ -1668,15 +1650,27 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     });
   }
 
+  get disableTableMenuButton(): boolean {
+    return !!(this.permissions && this.permissions.menu && this.permissions.menu.enabled === false);
+  }
+
   get showTableMenuButton(): boolean {
+    const permissionHidden = !!(this.permissions && this.permissions.menu && this.permissions.menu.visible === false);
+    if (permissionHidden) {
+      return false;
+    }
     const staticOpt = this.selectAllCheckbox || this.exportButton || this.columnsVisibilityButton || this.oTableColumnsFilterComponent !== undefined;
     return staticOpt || this.tableOptions.length > 0;
   }
 
   setOTableInsertableRow(tableInsertableRow: OTableInsertableRowComponent) {
-    this.oTableInsertableRowComponent = tableInsertableRow;
-    this.showFirstInsertableRow = this.oTableInsertableRowComponent.isFirstRow();
-    this.showLastInsertableRow = !this.showFirstInsertableRow;
+    const insertPerm: OPermissions = this.getActionPermissions(PermissionsUtils.ACTION_INSERT);
+    if (insertPerm.visible) {
+      tableInsertableRow.enabled = insertPerm.enabled;
+      this.oTableInsertableRowComponent = tableInsertableRow;
+      this.showFirstInsertableRow = this.oTableInsertableRowComponent.isFirstRow();
+      this.showLastInsertableRow = !this.showFirstInsertableRow;
+    }
   }
 
   clearSelectionAndEditing() {
@@ -1789,6 +1783,9 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   insertRecord(recordData: any, sqlTypes?: Object): Observable<any> {
+    if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_INSERT)) {
+      return undefined;
+    }
     if (!Util.isDefined(sqlTypes)) {
       let allSqlTypes = this.getSqlTypes();
       let sqlTypes = {};
@@ -1800,6 +1797,9 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   updateRecord(filter: any, updateData: any, sqlTypes?: Object): Observable<any> {
+    if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_UPDATE)) {
+      return Observable.of(this.dataSource.data);
+    }
     let sqlTypesArg = sqlTypes || {};
     if (!Util.isDefined(sqlTypes)) {
       let allSqlTypes = this.getSqlTypes();
@@ -1861,13 +1861,13 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return !Util.isDefined(value) || ((typeof value === 'string') && !value);
   }
 
-  protected setFiltersConfiguration(conf: any) {
+  setFiltersConfiguration(conf: any) {
     this.filterCaseSensitive = conf.hasOwnProperty('filter-case-sensitive') ? conf['filter-case-sensitive'] : this.filterCaseSensitive;
 
     const storedColumnFilters = this.oTableStorage.getStoredColumnsFilters(conf);
     this.showFilterByColumnIcon = storedColumnFilters.length > 0;
-    if (this.columnFilterOption) {
-      this.columnFilterOption.active = this.showFilterByColumnIcon;
+    if (this.oTableMenu && this.oTableMenu.columnFilterOption) {
+      this.oTableMenu.columnFilterOption.active = this.showFilterByColumnIcon;
     }
 
     if (this.oTableColumnsFilterComponent) {
@@ -1891,42 +1891,19 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     }
   }
 
-  onStoreConfigurationClicked(): void {
-    let dialogRef = this.dialog.open(OTableStoreConfigurationDialogComponent, {
-      width: '30vw',
-      disableClose: true
-    });
-    const self = this;
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        const configurationData = dialogRef.componentInstance.getConfigurationAttributes();
-        const tableProperties = dialogRef.componentInstance.getSelectedTableProperties();
-        self.oTableStorage.storeConfiguration(configurationData, tableProperties);
-      }
-    });
+  onStoreConfigurationClicked() {
+    if (this.oTableMenu) {
+      this.oTableMenu.onStoreConfigurationClicked();
+    }
   }
 
-  onApplyConfigurationClicked(): void {
-    let dialogRef = this.dialog.open(OTableApplyConfigurationDialogComponent, {
-      data: this.oTableStorage.getStoredConfigurations(),
-      width: '30vw',
-      disableClose: true
-    });
-    const self = this;
-    dialogRef.componentInstance.onDelete.subscribe(configurationName => this.oTableStorage.deleteStoredConfiguration(configurationName));
-    dialogRef.afterClosed().subscribe(result => {
-      if (result && dialogRef.componentInstance.isDefaultConfigurationSelected()) {
-        self.applyDefaultConfiguration();
-      } else if (result) {
-        let selectedConfigurationName: string = dialogRef.componentInstance.getSelectedConfigurationName();
-        if (selectedConfigurationName) {
-          self.applyConfiguration(selectedConfigurationName);
-        }
-      }
-    });
+  onApplyConfigurationClicked() {
+    if (this.oTableMenu) {
+      this.oTableMenu.onApplyConfigurationClicked();
+    }
   }
 
-  protected applyDefaultConfiguration() {
+  applyDefaultConfiguration() {
     this.oTableStorage.reset();
     this.initializeParams();
     this.parseVisibleColumns();
@@ -1937,7 +1914,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     this.reloadData();
   }
 
-  protected applyConfiguration(configurationName: string) {
+  applyConfiguration(configurationName: string) {
     let storedConfiguration = this.oTableStorage.getStoredConfiguration(configurationName);
     if (storedConfiguration) {
       const properties = storedConfiguration[OTableStorage.STORED_PROPERTIES_KEY] || [];
@@ -2074,6 +2051,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return height;
   }
 
+
   isDetailMode(): boolean {
     return this.detailMode !== Codes.DETAIL_MODE_NONE;
   }
@@ -2087,6 +2065,20 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     Util.copyToClipboard(JSON.stringify(selectedItems));
   }
 
+
+  viewDetail(item: any): void {
+    if (!this.checkEnabledActionPermission('detail')) {
+      return;
+    }
+    super.viewDetail(item);
+  }
+
+  editDetail(item: any): void {
+    if (!this.checkEnabledActionPermission('edit')) {
+      return;
+    }
+    super.editDetail(item);
+  }
 
 }